ECP-8476 - Update open invoice method to allow for discounts and create new openinvoice for checkout, capture and refund builder.
