+++ Gateway/Request/RefundDataBuilder.php
@@ -13,6 +13,7 @@ namespace Adyen\Payment\Gateway\Request;
 
 use Adyen\Payment\Helper\ChargedCurrency;
 use Adyen\Payment\Helper\Data;
+use Adyen\Payment\Helper\OpenInvoice;
 use Adyen\Payment\Model\ResourceModel\Invoice\CollectionFactory;
 use Adyen\Payment\Model\ResourceModel\Order\Payment\CollectionFactory as PaymentCollectionFactory;
 use Magento\Payment\Gateway\Data\PaymentDataObject;
@@ -51,23 +52,31 @@ class RefundDataBuilder implements BuilderInterface
     private $chargedCurrency;
 
     /**
+     * @var OpenInvoice
+     */
+    protected $openInvoiceHelper;
+
+    /**
      * RefundDataBuilder constructor.
      *
      * @param Data $adyenHelper
      * @param PaymentCollectionFactory $orderPaymentCollectionFactory
      * @param CollectionFactory $adyenInvoiceCollectionFactory
      * @param ChargedCurrency $chargedCurrency
+     * @param OpenInvoice $openInvoiceHelper
      */
     public function __construct(
         Data $adyenHelper,
         PaymentCollectionFactory   $orderPaymentCollectionFactory,
         CollectionFactory          $adyenInvoiceCollectionFactory,
-        ChargedCurrency            $chargedCurrency
+        ChargedCurrency            $chargedCurrency,
+        OpenInvoice                $openInvoiceHelper
     ) {
         $this->adyenHelper = $adyenHelper;
         $this->orderPaymentCollectionFactory = $orderPaymentCollectionFactory;
         $this->adyenInvoiceCollectionFactory = $adyenInvoiceCollectionFactory;
         $this->chargedCurrency = $chargedCurrency;
+        $this->openInvoiceHelper = $openInvoiceHelper;
     }
 
     /**
@@ -175,7 +184,7 @@ class RefundDataBuilder implements BuilderInterface
             );
 
             if ($this->adyenHelper->isPaymentMethodOpenInvoiceMethod($brandCode)) {
-                $openInvoiceFields = $this->getOpenInvoiceData($payment);
+                $openInvoiceFields = $this->openInvoiceHelper->getOpenInvoiceData($payment);
 
                 //There is only one payment, so we add the fields to the first(and only) result
                 $requestBody[0]["additionalData"] = $openInvoiceFields;
@@ -193,91 +202,4 @@ class RefundDataBuilder implements BuilderInterface
         return $request;
     }
 
-    /**
-     * @param InfoInterface $payment
-     * @return array|mixed
-     */
-    protected function getOpenInvoiceData($payment)
-    {
-        $formFields = [];
-        $count = 0;
-
-        // Construct AdyenAmountCurrency from creditmemo
-        /**
-         * @var Creditmemo $creditMemo
-         */
-        $creditMemo = $payment->getCreditMemo();
-
-        foreach ($creditMemo->getItems() as $refundItem) {
-            $numberOfItems = (int)$refundItem->getQty();
-            if ($numberOfItems == 0) {
-                continue;
-            }
-
-            ++$count;
-            $itemAmountCurrency = $this->chargedCurrency->getCreditMemoItemAmountCurrency($refundItem);
-
-            $formFields = $this->adyenHelper->createOpenInvoiceLineItem(
-                $formFields,
-                $count,
-                $refundItem->getName(),
-                $itemAmountCurrency->getAmount(),
-                $itemAmountCurrency->getCurrencyCode(),
-                $itemAmountCurrency->getTaxAmount(),
-                $itemAmountCurrency->getAmount() + $itemAmountCurrency->getTaxAmount(),
-                $refundItem->getOrderItem()->getTaxPercent(),
-                $numberOfItems,
-                $payment,
-                $refundItem->getId()
-            );
-        }
-
-        // Shipping cost
-        $shippingAmountCurrency = $this->chargedCurrency->getCreditMemoShippingAmountCurrency($creditMemo);
-        if ($shippingAmountCurrency->getAmount() > 0) {
-            ++$count;
-            $formFields = $this->adyenHelper->createOpenInvoiceLineShipping(
-                $formFields,
-                $count,
-                $payment->getOrder(),
-                $shippingAmountCurrency->getAmount(),
-                $shippingAmountCurrency->getTaxAmount(),
-                $shippingAmountCurrency->getCurrencyCode(),
-                $payment
-            );
-        }
-
-        // Adjustment
-        $adjustmentAmountCurrency = $this->chargedCurrency->getCreditMemoAdjustmentAmountCurrency($creditMemo);
-        if ($adjustmentAmountCurrency->getAmount() != 0) {
-            $positive = $adjustmentAmountCurrency->getAmount() > 0 ? 'Positive' : '';
-            $negative = $adjustmentAmountCurrency->getAmount() < 0 ? 'Negative' : '';
-            $description = "Adjustment - " . implode(' | ', array_filter([$positive, $negative]));
-
-            ++$count;
-            $formFields = $this->adyenHelper->createOpenInvoiceLineAdjustment(
-                $formFields,
-                $count,
-                $description,
-                $adjustmentAmountCurrency->getAmount(),
-                $adjustmentAmountCurrency->getCurrencyCode(),
-                $payment
-            );
-        }
-
-        $formFields['openinvoicedata.numberOfLines'] = $count;
-
-        //Retrieve acquirerReference from the adyen_invoice
-        $invoiceId = $creditMemo->getInvoice()->getId();
-        $invoices = $this->adyenInvoiceCollectionFactory->create()
-            ->addFieldToFilter('invoice_id', $invoiceId);
-
-        $invoice = $invoices->getFirstItem();
-
-        if ($invoice) {
-            $formFields['acquirerReference'] = $invoice->getAcquirerReference();
-        }
-
-        return $formFields;
-    }
 }
